+++ simplexss/core/transports/websocket/__init__.py
@@ -1,5 +1,7 @@
 from .services import WebsocketService
+from .sessions import WebsocketSession
 
 __all__ = [
     'WebsocketService',
+    'WebsocketSession',
 ]